Extraced DataGetter to a new file
